fitness-bits-bobs/matrix-fitness-bridge/requirements-enhanced.txt
matrix-nio[e2e]==0.20.2
aiohttp==3.8.6
requests==2.31.0
python-dotenv==1.0.0
asyncio-mqtt==0.11.1
